+++ csharp.xUnit/GildedRose/Strategies/AgedBrieUpdater.cs
@@ -1,4 +1,4 @@
-namespace GildedRoseKata;
+namespace GildedRoseKata.Strategies;
 
 public class AgedBrieUpdater : ItemUpdaterBase
 {
