+++ csharp.xUnit/GildedRose/Strategies/NormalItemUpdater.cs
@@ -1,4 +1,4 @@
-namespace GildedRoseKata;
+namespace GildedRoseKata.Strategies;
 
 public class NormalItemUpdater : ItemUpdaterBase
 {
