+++ csharp.xUnit/GildedRose/Strategies/SulfurasUpdater.cs
@@ -1,4 +1,4 @@
-namespace GildedRoseKata;
+namespace GildedRoseKata.Strategies;
 
 public class SulfurasUpdater : ItemUpdaterBase
 {
